+++ src/client/net/commands.cpp
@@ -45,7 +45,7 @@ bool isEmptyImage(const QImage &image)
 // Recursively split image into small enough pieces.
 // When mode is anything else than MODE_REPLACE, PutImage calls
 // are expensive, so we want to split the image into as few pieces as possible.
-void splitImage(int ctxid, int layer, int x, int y, const QImage &image, int mode, bool skipempty, QList<protocol::MessagePtr> &list)
+void splitImage(uint8_t ctxid, uint16_t layer, int x, int y, const QImage &image, uint8_t mode, bool skipempty, QList<protocol::MessagePtr> &list)
 {
 	Q_ASSERT(image.format() == QImage::Format_ARGB32_Premultiplied);
 
@@ -140,7 +140,7 @@ MessagePtr terminateSession()
 	return serverCommand("kill-session");
 }
 
-QList<protocol::MessagePtr> putQImage(int ctxid, int layer, int x, int y, QImage image, paintcore::BlendMode::Mode mode, bool skipempty)
+QList<protocol::MessagePtr> putQImage(uint8_t ctxid, uint16_t layer, int x, int y, QImage image, paintcore::BlendMode::Mode mode, bool skipempty)
 {
 	QList<protocol::MessagePtr> list;
 
@@ -177,7 +177,7 @@ QList<protocol::MessagePtr> putQImage(int ctxid, int layer, int x, int y, QImage
 	return list;
 }
 
-protocol::MessagePtr setCanvasBackground(int ctx, const QColor &color)
+protocol::MessagePtr setCanvasBackground(uint8_t ctx, const QColor &color)
 {
 	return protocol::MessagePtr(new protocol::CanvasBackground(ctx, color.rgba()));
 }
