+++ src/client/net/commands.h
@@ -98,10 +98,10 @@ protocol::MessagePtr terminateSession();
  * Note: when using MODE_REPLACE, the skipempty parameter should be set to false, except when you know
  * the layer is blank.
  */
-QList<protocol::MessagePtr> putQImage(int ctxid, int layer, int x, int y, QImage image, paintcore::BlendMode::Mode mode, bool skipempty=true);
+QList<protocol::MessagePtr> putQImage(uint8_t ctxid, uint16_t layer, int x, int y, QImage image, paintcore::BlendMode::Mode mode, bool skipempty=true);
 
 //! Set the canvas background
-protocol::MessagePtr setCanvasBackground(int ctxid, const QColor &color);
+protocol::MessagePtr setCanvasBackground(uint8_t ctxid, const QColor &color);
 
 }
 }
